--- packages/drawer/src/EmbeddedDrawerLayout/EmbeddedDrawerLayout.types.ts
@@ -1,3 +1,0 @@
-import { LayoutComponentProps } from '../LayoutComponent/LayoutComponent.types';
-
-export type EmbeddedDrawerLayoutProps = Omit<LayoutComponentProps, 'drawer'>;
