+++ packages/drawer/src/LayoutComponent/LayoutComponent.tsx
@@ -6,8 +6,8 @@ import LeafyGreenProvider, {
 
 import { DisplayMode } from '../Drawer';
 import { useDrawerLayoutContext } from '../DrawerLayout';
-import { EmbeddedDrawerLayout } from '../EmbeddedDrawerLayout';
-import { OverlayDrawerLayout } from '../OverlayDrawerLayout';
+import { EmbeddedDrawerLayout } from './EmbeddedDrawerLayout';
+import { OverlayDrawerLayout } from './OverlayDrawerLayout';
 
 import { LayoutComponentProps } from './LayoutComponent.types';
