+++ packages/drawer/src/LayoutComponent/OverlayDrawerLayout/OverlayDrawerLayout.types.ts
@@ -1,0 +1,3 @@
+import { LayoutComponentProps } from '../../LayoutComponent/LayoutComponent.types';
+
+export type OverlayDrawerLayoutProps = Omit<LayoutComponentProps, 'drawer'>;
