--- packages/drawer/src/OverlayDrawerLayout/OverlayDrawerLayout.types.ts
@@ -1,3 +1,0 @@
-import { LayoutComponentProps } from '../LayoutComponent/LayoutComponent.types';
-
-export type OverlayDrawerLayoutProps = Omit<LayoutComponentProps, 'drawer'>;
